sockets: polling one message at a time
When the first couple messages are received in a single poll, the last a few
messages' addr field may be set to FI_ADDR_NOTAVAIL since the
connection table has not been established yet -- the connection message is in
the same batch yet to be processed. Message with FI_ADDR_NOTAVAIL will
have trouble matching as it skips the checking of source, resulting in
message mismatch.

This patch tries to fix it by limit number of polled message to 1.

Signed-off-by: Hui Zhou <hzhou321@anl.gov>