fabtests: Add --with-cuda configure argument
The --with-cuda configure argument adds CUDA runtime support to
fabtests. If a test application is integrate with CUDA support, the
application must dlopen CUDA runtime. The reason for this requirement is
to allow for a single fabtests build to operator with and without CUDA
runtime being present on a node.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>