prov/tcp: Read from socket if staging buffer is empty
In tcpx_recv_hdr and tcpx_recv_msg_data, we either read from the
staging buffer if data is available, or from the socket if it is
empty.  If we empty the staging buffer, it's likely that data is
available to be read from the socket.  However, we require looping
back through the progress loop (usually returning to the app)
before checking the socket.

Update the code to immediately try reading from the socket once
the staging buffer is empty.

As part of this change, convert tcpx_recv_hdr to return ssize_t,
rather than int, since it returns the number of bytes read in the
success case.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>