prov/cxi: Refactor collectives for one PTE per MC addr per EP.
The original design called for one PTE that would manage all MC
addresses for a given endpoint, using different RMU indices to map the
different MC addresses to the same PTE index.

The underlying CXI driver, after discussion, ended with one PTE
per MC address (per EP), so the PTEs needed to be distributed into
separate multicast objects.

Eight reduction objects are added to the MC object, to separately track
state for each of the eight possible concurrent reductions.

This now requires a join_collective() implementation to create the
individual MC objects, which led to expansion of the test environment.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>