fabtests/common: save common mr_desc in context array
The context array allows separation of all potentially concurrent
transfers to have separate fi_contexts for use in each transfer
as well as (optional) separated memory buffer allocation and
registration.

The option FT_OPT_ALLOC_MULT_MR allows for allocation and registration
of the separate buffers. If FT_MR_OPT_ALLOC_MULT_MR is not set, the
context array will still be allocated, but the common/single memory
allocation and registration will be used instead.

The initialization of the context array loops through all the contexts
and either allocates and registers the buf/mr and gets the mr desc for
each buffer if FT_OPT_ALLOC_MULT_MR is set or points it to the common/
single memory buffer.

Multiple tests, such as the functions/rdm_tagged_peek test, use the
context array for using different fi_contexts in multiple outstanding
operations, but use the common buffers and mr_desc.

This change doesn't fix any current usage (since tests like the tagged
peek test use the array[i].context but not the array[i].mr_desc and
instead use the global mr_desc) but makes the ctx_array usage
consistent so that a test could use the buf, mr, and desc for a
specific array entry and everything would match.

Signed-off-by: aingerson <alexia.ingerson@intel.com>