prov/efa: fix a bug in rxr_rma_readmsg()
Currently in rxr_rma_readmsg(), when error was encountered,
the function return immediately. This is wrong because a lock
on rxr_ep->util_ep.lock has been acquired at the beginning
of the function and was not released.

This patch fix the issue by set err then "goto out", so that
the lock will be released.

Signed-off-by: Wei Zhang <wzam@amazon.com>