prov/efa: Completion handling for RX with prefix
From the completion context, determine if the RX buffer belongs to the
application or the provider and handle it appropriately. When handling a
user buffer, the reordering and unexpected message handling are
circumvented, as is the staging in those corresponding pkt_entry pools.
This path is only supported for RXR_EAGER_MSGRTM_PKT, so a completion is
immediately written to the application.

Signed-off-by: Raghu Raja <craghun@amazon.com>