core/abi: Fix ABI 1.1 EP attribute definition
The current fi_ep_attr structure was defined for the 1.1 ABI
change, not 1.2.  struct fi_info_1_1 incorrectly references
the wrong ep attribute structure.  It should reference the
existing fi_ep_attr.

Adding to confusion, struct fi_ep_attr_1_2 is the same as
fi_ep_attr.  However, the structure should have been called
'1_1' instead of '1_2'.  But given that it's the same as the
external structure, we don't need the definition.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>