Move symver check out of main() in configure check.
System: Summit
Version: master branch (as of 7/23/20)

Using the XL compiler, libfabric is able to configure, but fails on make:
```
  CC       src/src_libfabric_la-fabric.lo
src/fabric.c:790:1: error: 1540-2991 The expression is not supported.
CURRENT_SYMVER(fi_freeinfo_, fi_freeinfo);
^
./include/ofi_abi.h:130:2: note: expanded from macro 'CURRENT_SYMVER'
        asm(".symver " #name "," #api "@@" CURRENT_ABI)
        ^
1 error generated.
Error while processing src/fabric.c.
```

This is because xlc does not support asm statements outside of a function body, but the test of symver:
```
AC_TRY_LINK([],
        [__asm__(".symver main_, main@ABIVER_1.0");],
        [
                AC_MSG_RESULT(yes)
                ac_asm_symver_support=1
        ],
        [
                AC_MSG_RESULT(no)
                ac_asm_symver_support=0
        ])
```

Inserts the asm call in `main()` in the conftest.

Signed-off-by: Philip Davis <philipdavis01@gmail.com>