fabtests: Add --with-rocr configure argument
The --with-rocr configure argument adds ROCR runtime support to
fabtests. If a test application is integrate with ROCR support, the
application must dlopen ROCR runtime. The reason for this requirement is
to allow for a single fabtests build to operator with and without ROCR
runtime being present on a node.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>