prov/efa: Update FI_MSG_PREFIX mode and attributes
For applications that can provide prefix buffers with FI_MSG, advertise
the correct msg_prefix_size. Based on the mode bits requested by the
application, determine if the endpoint can support zero-copy receives or
not.

Given the prefix size can not adapt to the protocol or the REQ packet's
header size after endpoint initialization, this commit also forces a
handshake before any user data is sent.

Signed-off-by: Raghu Raja <craghun@amazon.com>