prov/efa: move rdma-core related initialization to init_lower_efa_prov()
Currently, the efa provider will register atfork callback function and
set the environment variable RDMAV_HUGEPAGES_SAFE in efa_fabric().

The intention is that if an application needs fork support, it should
set the environment variable IBV_FORK_SAFE or RDMAV_FORK_SAFE to enable
it. Because EFA provider uses huge page, the provider should set
environment variable RDMAV_HUGEPAGES_SAFE, so when the fork support
was enabled, it also support huge page.

There is an issue with current implementation: RDMAV_HUGEPAGES_SAFE
was set too late.

For it to be effective, RDMAV_HUGEPAGE_SAFE must be set before the function
ibv_fork_init() was called. However, the EFA provider call ibv_fork_init()
during the initialization of the provider, the call stack is:

   init_lower_efa_prov()->efa_init_info()->efa_device_init()
   -> ibv_get_device_list() -> ibverbs_init() -> ibv_fork_init()

which is before efa_fabric() was called.

To address this issue, this patch moves the registration of atfork callback
function and the setting of RDMAV_HUGEPAGES_SAFE to the beginning of
init_lower_efa_prov(), before the call of efa_init_info().

Accordingly, this patch moves the unset of RDMAV_HUGEPAGES_SAFE from
efa_fabric_close() to fi_efa_fini().

Signed-off-by: Wei Zhang <wzam@amazon.com>