prov/cxi: Return EIO for a xfer canceled by PCT
Do not confuse a canceled Receive requests with a request that was canceled by
PCT. Add a new 'unlinked' field to a Receive request to indicate whether it was
canceled by a libfabric user. If the request RC is C_RC_CANCELED it can then
only mean only one thing: PCT canceled the request.

Signed-off-by: Zach Tiffany <zachary.tiffany@gmail.com>