prov/psm2: Treat dynamic connection errors as fatal
The function psmx2_epid_to_epaddr() is the core for dynamic connection
setup. Previous it returned error code to the caller when connection
failed. However, the call sites simply used assert() to catch the
errors because trying to deliver the errors further to upper layers
would introduce extra overhead to the critical path of communication.
Since assert() is optimized out for non-debug build, this means the
errors are unhandled in production code.

Change psm2_epid_to_epaddr() to abort on connection errors. The log
message is enhanced with suggestion for handling connection timeout.
As the side effect, the call sites become less cluttered.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>