Initial changes to reduce number of files to review.
Changes generall have no effect on function, but reduce the number of
files to review in subsequent commits, and simplifies diffs due to name
changes.

This also tests for send/recv allowed in EP before attempting to enable
the collectives, since this code tries to create a PTE for write-only
endpoints in test rma_opt/opt_write.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>