prov/tcp: Fix deadlock due to AB-BA lock ordering
The eq->close_lock and ep->lock may be acquired in different orders.
The ep->lock protects against the state of the ep.  While the
eq->close_lock is intended to protect against processing an EP's
connection state while an application is trying to close the EP.

process_cm_ctx holds the close_lock.  In both server_send_cm_accept
and client_recv_connresp paths, ep_enable_xfers will acquire the ep
lock.  This is a normal flow that sets up the connection.

If an error occurs on a socket, for example the peer has disconnected,
then several paths will attempt to shutdown the socket.  For example,
process_tx_entry, process_rx_entry, remote_write, remote_read, and
progress_rx can all call ep_shutdown_report.  That path will call
ep_wait_fd_del, which acquires the close_lock.

The purpose of having the close_lock is to avoid the situation where
an application thread reads the EQ, finds an event for an EP, and
tries to process it.  This is really only an issue if the application
tries to close the endpoint, which frees it.  In the case where
ep_shutdown_report is called as a result of an error occurring on the
socket, this protection isn't needed.  A simple state check that the
EP is still enabled is sufficient.

However, we still need to sync freeing the endpoint with any thread
processing the EQ.  To fix this issue, we still need to sync ep_close
with process_cm_ctx.  It's okay to acquire the close_lock from ep_close,
since the ep lock is not held.

Move the close_lock out of ep_wait_fd_del, to fix the deadlock issue,
and add back to ep_close to provide the proper synchronization.  Add
checks in ep_close to verify that the EP has been bound to an EQ prior
to accessing the EQ.  The latter handles the case where an EP is
created and destroyed without being used.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>