prov/tcp: Rework CM state machine to fix lock inversion
The tcp provider has an AB-BA and ABA lock inversion problem.

thread 1:
util_wait_fd_try - wait->lock
-> tcpd_try_func - ep->lock

thread 2:
ofi_cq_readfrom
-> tcpx_cq_progress - ep->lock
--> tcpx_progress_rx
---> tcpx_ep_shutdown_report
----> ep_wait_fd_del - wait->lock

The above shows the AB-BA deadlock.  Additionally, util_wait_fd_try can call
through to ofi_cq_readfrom.  That can result in acquiring the wait->lock
recursively.

The core issue is how the tcp provider handles remote disconnects, which can
occur at anypoint, even during connection setup.  To address this problem,
we rework the tcp CM state machine.  The states are updated to reflect the
full range of states that the endpoint may be in.  If an error occurs in
the connection process, or if we detect that the peer has disconnected
(indicated, for example, if socket recv returns 0), then the endpoint is
moved into a disconnected state.

A disconnected endpoint is disabled.  When disabling an endpoint, its
socket is removed from any corresponding epoll fd or pollfds.  This
prevents the socket from generating events on the epoll fd or pollfds,
which avoid application threads from looping on that event repeatedly.
However, the endpoint maintains its relationship with the wait set itself.
This avoids needing to acquire the wait->lock while holding the ep->lock.
That in turn avoids both locking issues.

When transitioning through the CM states, we restructure code in places to
simplify error handling and provide consistent meaning to each CM state.

The endpoint is not removed from referencing the wait sets until it is
destroyed by the application.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>