prov/efa: fix a bug in rxr_pkt_proc_matched_medium_rtm()
rxr_pkt_proc_matched_medium_rtm() call

   rxr_pkt_req_hdr_size()

in the loop to get header size of each packet entry in a linked list.

Currently, the input to that function is "pkt_entry" which is wrong,
because it is the first packet entry. The input should be "cur",
which is the current packet entry.

This patch fix the issue.

Signed-off-by: Wei Zhang <wzam@amazon.com>