prov/cxi: increase control EQ size
On receipt of the first packet of a message the LPE must reserve enough
event slots in the EQ for any events that may need to be generated in
the course of processing the message.

MR PtlTEs use the control EQ. Increase the size of the control EQ to
avoid 'EQ full' errors due to many outstanding LPE EQ reservations.

Use 16k for now. This is space for 256 events. This may need refinement
if LPE can process more than 256 messages at once.

For NETCASSINI-2195.

Signed-off-by: Zach Tiffany <zachary.tiffany@hpe.com>