prov/efa: Do not release rx_entry of EOR packet in EAGAIN case
Currently, we release the rx_entry after the receiver posts EOR packet
and it's failed with FI_EAGAIN, which results in that the EOR packet
does not get re-posted later, and the sender can not write send
completion accordingly.

This patch fixes the above issue by only releasing rx_entry of EOR
packet in FI_SUCCESS case.

Signed-off-by: Jie Zhang <zhngaj@amazon.com>