util/mem_hooks: port Open MPI overwrite patcher code
The previous implementation (from Open MPI patcher/linux) does not
intercept syscalls made by glibc. Port the overwrite patcher
implementation from Open MPI. This rewrites the beginning of glibc's
function call with a jump to our intercept and our intercept will make
the syscall directly.

Only the x86_64 and aarch64 support was ported over and we're only
supporting this on Linux at this time.

Also, remove dlopen and sbrk hooks. dlopen was needed to update the
pointers when another shared library is loaded and sbrk is a wrapper for
brk.

See issue #6244

Signed-off-by: Robert Wespetal <wesper@amazon.com>