core/net: Mark if addresses support local/remote communication
When collecting the list of addresses available in the system, also
record if the address supports local and/or remote communication.

When fi_getinfo is called, a list of fi_info structures is created
based on the provider's base fi_info list.  That list may be expanded
in order to report all addresses that are usable to the app.  For
example, see udpx_getinfo -> ofi_ip_getinfo -> util_getinfo_ifs.

The list of available addresses in the system is created for each
provider through ofi_get_list_of_addr.  When constructing that list,
the loopback addresses are appended to the end of the list.

Currently, regardless of the application request, one fi_info
is created for each address and returned to the user.  This can
result in returnin a loopback address to the app even though it
requests FI_REMOTE_COMM.

Expand the address entry (ofi_addr_list_entry) to record if the
address supports FI_REMOTE_COMM / FI_LOCAL_COMM.  Use that to
verify if the address is usable prior to returning it to the user,
and update the COMM flags as needed.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>