prov/sockets: Retry address lookup if not set
If several messages are received immediately after a connection has been
established, the AV index for the connection may not be set when new
Rx process entries are initialized.  The result is that those messages may
be associated with the wrong data buffer due to the address being set to
FI_ADDR_UNAVAIL.

To handle this case, see if the pe_entry address has been set when handling
the receive.  This allows the first message to be processed completely, which
can set the av_index.  Other received messages will then have the av_index
set correctly.

This problem has shown up on BSD systems.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>