prov/efa: us array for iov and rma_iov in rxr_read_entry
Currently, iov and rma_iov in rxr_read_entry is a pointer pointing
to iov and rma_iov in rx_entry and tx_entry.

This patch make them arrays, and initialize the content by memcpy.
This is to prepare for an upcoming change, for which the iov and
rma_iov has to be constructed.

Signed-off-by: Wei Zhang <wzam@amazon.com>