prov/efa: fix a bug that return of rxr_pkt_proc_data() is ignored
Currently, return type of rxr_pkt_proc_data() int. If it encoutered
error, the function will return error code.

However, the caller of this function is not checking return of
rxr_pkt_proc_data().

This patch fix the issue by changing the return type of this function
to void, and have it write error completion entry if error has
been encountered

Signed-off-by: Wei Zhang <wzam@amazon.com>