prov/efa: introduce bytes_received and bytes_copied to rx_entry
Currently, rx_entry->bytes_done was used to track data that was
both received and copied to receiving buffer.

This patch introduced two counters: bytes_received and bytes_copied to
track data in each state. rx_entry->bytes_done was removed.

This is for an upcoming change, in which data received is not immediately
copied to receiving buffer.

Other changs are:

1. RX completion is written when byte_copied reached rx_entry->total_len.

2. for long message protocl, receiving window is calculated according
   to bytes_received. Update of ep->rx_pending_list also depends on
   bytes_received.

3. for medium message protocol, the update of msg_id->rx_entry also
   depends on bytes_received.

Signed-off-by: Wei Zhang <wzam@amazon.com>