hmem: Remove cuInit() driver initialization
PR #6160 introduced the driver initialization as a sanity check to to
see if a CUDA-capable GPU is available before allowing FI_HMEM
operations. We use the CUDA runtime API for our CUDA needs in libfabric,
and should leave the driver initialization up to the runtime to manage.
We do not use a separate context for libfabric's needs either.

Moreoever, cuInit() forks and execs `nvidia-modrobe` to setup device files
and load drivers if they are missing. Rdma-core based providers are
not fork-safe by default, and this implicit fork causes the EFA provider
to abort for all applications on systems without a NVIDIA GPU when built
with CUDA.

Signed-off-by: Raghu Raja <craghun@amazon.com>