util: Add a memory monitor read-write list lock
Due to possibly needing to iterate over a monitor list while releasing
the mm_lock, the mm_lock is insufficient to protect thread safety. Thus
we need a separate lock for the list iteration. Due to a possible
recursion when iterating over the lists, we need to have this be a
read-write lock prioritizing read locks. Thus, the write lock is only
acquired through a nonblocking loop.

Signed-off-by: William Zhang <wilzhang@amazon.com>