prov/shm: fix shm mapping/addressing issues
Fixes an issue where if a peer sends a CMA message to a peer
which does not have the sender inserted into its AV, shm seg faults.
This is because CMA messages require the target to have the sender
mapped which currently only happens on av_insert.

To fix this, add a handshake on the first send to ensure any target
peers have the sender mapped before processing incoming messages.

To do this, the patch does the following:
- Adds ofi_ctrl_* offset definition for ctrl ops and new ops
if needed in the future
- Adds a ofi_ctrl_connreq message that sends the sender information
to the receiver before sending any messages
- Use internal shm_addr indeces instead of using the application's
fi_addr in order to map non-inserted peers
- Have smr_map_add handle duplicate calls and set/return the
shm_addr to use
- Have verify_peer lookup the fi_addr and return the shm_addr to use
or FI_ADDR_UNSPEC if the mapping exchange has not finished instead
of using the fi_addr idrectly as the shm index
- Change util_av to store shm_addr (fi_addr/index) instead of name
- Split uint64_t sar status into uint32_t sar_status and uint32_t
name_sent

Signed-off-by: aingerson <alexia.ingerson@intel.com>