prov/shm: change shm_id to int64
Change storage and usage of shm_id to int64_t,
not fi_addr_t since they are not interchangeable
anymore. Also rename some variables to make it clear
that the shm index is not an address.
This helps clarify and separate the two implementations.

As a result of the cleaning up and separation, several
bugs are also fixed in this patch where fiaddrs were
incorrectly being used as shm indeces
- on av_remove, lookup fiaddr and remove by index, not fiaddr
- on av_lookup lookup fiaddr and get peer_smr by index, not fiaddr
- store shm index in receive entries to match with incoming
  peer shm indeces in command headers
- complete rx entries with fiaddr for source address reporting,
  not the cmd addr which is a shm index
     - add fiaddr field in smr_map to lookup fiaddr from shm_id

Signed-off-by: aingerson <alexia.ingerson@intel.com>