src/hmem_ze: add p2p enabled function
Some earlier versions of L0 do not support transfers
across different devices. Because of this limitation,
they cannot enable the IPC protocol and instead must
go through the SAR protocol. To allow the shm provider
to enable or disable the IPC protocol, expose an
L0-specific function to check for the limitation. The
new functions checks for the limitation between all
combinations of devices during ZE initialization to
verify all IPC transfers are possible.

Signed-off-by: aingerson <alexia.ingerson@intel.com>