prov/rxd: fix bug preventing segmentation acks
Data packets that are at the end of a window should
trigger an ACK to the sender to drive progress on
the sender. However, due to incorrect parentheses, the
ACK was always getting skipped, requiring a resend in
order for an ACK to get sent. Fixing this significantly
improves performance at larger messages.

Signed-off-by: aingerson <alexia.ingerson@intel.com>