prov/rxm: Allow re-connecting to peers
If we fail to establish a connection to a peer, or if after a
connection has been established, the peer shuts down, the
connection map entry that corresponds with the peer's address
in the AV is cleared.  The result is that any attempts to
send to the same address will fail with -EHOSTUNAVAIL.

In the case where a connection to the peer fails, this can
be the result of starting a client process prior to the
server process being started.  (The server will be using a
well-known address).  This problem occurs running DAOS
tests.

A similar problem can occur if a peer is re-started.  The
client's connection to the server can be shutdown.  Once
the server is restarted, the client should be able to
communicate again.  However, since the server address is the
same, all future attempts to send to the server will fail.

To fix this, we allocate a new cmap handle if one is not
available.  This resets the connection state, allowing the
formation of a new connection to the server, so that further
transfers are possible.  Additionally, if the connect request
fails immediate with ECONNREFUSED, we return EAGAIN, so that
the application will retry the request.  ECONNREFUSED can
occur when using the loopback address.

Includes some adjacent minor code cleanups to improve
readability.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>