configure: Rework FI_CHECK_PACKAGE[(cuda)] logic
The current logic was silently ignoring a failure to find the CUDA
runtime libraries. We would want the configure to abort if an explicit
request to build with CUDA fails to find the headers/libs in the
provided dir-path (or they are absent in the standard system paths
when no path is provided). When `--with-cuda` is not provided, we
still check to see if the runtime is available and define HAVE_LIBCUDA
if it is.

Signed-off-by: Raghu Raja <craghun@amazon.com>