prov/efa: fix a bug by removing address insertion during initialization
Currently, during end point initialization, the EP's own address is
inserted to address vector to get a libfabric address. The libfabric
address was later used to post a read to EP itself.

Inserting an address during initialization would cause the first
address inserted by application to be 1, which is against libfabric
standard.

The standard states that
'''
The index of the first address inserted into an FI_AV_TABLE will be 0,
and successive insertions will be given sequential indices.
'''

This patch address the issue by removing the insertion, and let
rxr_read_post() call efa_rma_post_read() directly.

efa_rma_post_read() has an argument self_comm, when it is true,
caller does not need to provide address, and the read will be posted to
end point itself.

Signed-off-by: Wei Zhang <wzam@amazon.com>