core/uffd: Add comment describing potential deadlock
This comment describes the behavior of the userfault monitor thread:

The userfault fd monitor requires for events that could
trigger it to be handled outside of the monitor functions
itself. When a fault occurs on a monitored region, the
faulting thread is put to sleep until the event is read
via the userfault file descriptor. If this fault occurs
within the userfault handling thread, no threads will
read this event and our threads cannot progress, resulting
in a hang.

Signed-off-by: William Zhang <wilzhang@amazon.com>