prov/efa: fix a bug in efa_get_gdr_support()
Currently, efa_get_gdr_support() call fi_read_file()
with pointer to a single char gdr_support, then call
atoi() on that pointer.

This is wrong because the expected input of atoi()
is an string terminated by '\0', but gdr_support is
not necessary followed by '\0' in stack.

This patch address the issue by defining a string
gdr_support_str, and call fi_read_file() to read file
to it, and calling strncmp() to compare it with "1".

If also make change to the caller. The caller will now
check for negative return value, and will not proceed
if return value is negative.

Signed-off-by: Wei Zhang <wzam@amazon.com>