prov/efa: Bugfix FI_MR_LOCAL only turns off cache for domain
Due to efa_mr_cache_enable being a global variable, it is checked for
every domain. The check for FI_MR_LOCAL to disable efa_mr_cache_enable
caused a bug where domains with different mr_mode settings would cause
the domains with caches to not cleanup their cache. Due to needing a
domain specific check, we now check if the cache was malloc'd or not and
do not change efa_mr_cache_enable.

Signed-off-by: William Zhang <wilzhang@amazon.com>