core/mr_cache: Remove storage abstraction and call rbtree directly
The MR cache defines OFI_MR_STORAGE_USER, which in theory allows
a provider to specify their own storage mechanism for maintaining
the MR cache.  However, it's never been used (at least in any
upstream code).  The result is that the code carries extra
complexity for a feature that is not used.

Remove the storage type for the mr cache and have it just be
the red-black tree, which is what it is anyway.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>