util/av: Replace av->count with ofi_av_size
The AV count reflects the original size of the AV, not the
current size.  The current size is now available through
ofi_av_size().  Remove av->count and update call sites to
use the ofi_av_size() function instead.

This fixes a bug where an AV set is created based on the
original size of the AV, rather than the current size.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>