util/av: Add function to return current max AV size
The AV may grow dynamically in order to handle apps that
cannot easily determine the proper size up front.  The
actual size of the AV is not av->count, but the size
of the buffer pool underneath the AV where the addresses
are stored.

Provide a function to obtain that value, which will be used
instead of reading the AV count.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>