dequeue rxc entry when an UNLINK event is received
NETCASSINI-2286

Mercury testing exposed an issue in the case where a rxc_entry is reused following a request cancellation.  The rxc_entry was not being removed from the queue following the cancellation/unlink, which eventually led to a seg fault when attempting to place the entry on the queue again in a subsequent test. With this change, the rxc_entry is removed when the (non-auto) UNLINK event is received.

Signed-off-by: Roger Connaughty <roger.connaughty@hpe.com>