prov/shm: add unix socket infrastructure for ZE
L0 requires a unix domain socket for IPC support. This
adds the initial infrastructure for creating, initializing,
and connecting domain sockets within the shm provider for
use later in the IPC protocol.

Summary of changes:
- Move definition of ZE_MAX_DEVICES for use in shm provider.
- Add smr flag to indicate whether the peer has a socket open for IPC.
- Add cmap entry for storing exchanged information with the peers. This
includes the device fds that eventually will be received from the peer.
- All unix socket paths will have the prefix /dev/shm/ze_ for simplicity
and cleanliness.
- Add listener thread to process incoming connection/exchange requests.
- Add epoll/signal for the EP to wake up and stop the listener thread.
- Each smr will have a unique, assigned sock_name defined as the following:
{PID}:{domain_id}:{ep_id}. The shm region name cannot be used
because a unix domain socket name has a limit of 108 characters while the
shm region name has a limit of 256 characters.
- Store ep_id in the smr_ep for socket name creation.
- To make sure only one connection is established per peer pair, the full name
for the socket will be the two peers' names in alphabetical order and
concatenated with '_'
- Add sock_name_list (similar to ep_name_list) to store all opened listener
sockets in order to unlink socket names on unclean exits.
- After the device fds are exchanged, the communicating socket is closed as it
only needs to be used once.

Signed-off-by: aingerson <alexia.ingerson@intel.com>